Apply clang-tidy

+++ source/fe/fe_system.cc
@@ -135,7 +135,7 @@ FESystem<dim, spacedim>::FESystem(const FiniteElement<dim, spacedim> &fe1,
                                                                   &fe2,
                                                                   n2),
       FETools::Compositing::compute_nonzero_components(&fe1, n1, &fe2, n2))
-  , base_elements((n1 > 0) + (n2 > 0))
+  , base_elements(static_cast<int>(n1 > 0) + static_cast<int>(n2 > 0))
 {
   std::vector<const FiniteElement<dim, spacedim> *> fes;
   fes.push_back(&fe1);
@@ -169,7 +169,8 @@ FESystem<dim, spacedim>::FESystem(const FiniteElement<dim, spacedim> &fe1,
                                                        n2,
                                                        &fe3,
                                                        n3))
-  , base_elements((n1 > 0) + (n2 > 0) + (n3 > 0))
+  , base_elements(static_cast<int>(n1 > 0) + static_cast<int>(n2 > 0) +
+                  static_cast<int>(n3 > 0))
 {
   std::vector<const FiniteElement<dim, spacedim> *> fes;
   fes.push_back(&fe1);
@@ -218,7 +219,8 @@ FESystem<dim, spacedim>::FESystem(const FiniteElement<dim, spacedim> &fe1,
                                                        n3,
                                                        &fe4,
                                                        n4))
-  , base_elements((n1 > 0) + (n2 > 0) + (n3 > 0) + (n4 > 0))
+  , base_elements(static_cast<int>(n1 > 0) + static_cast<int>(n2 > 0) +
+                  static_cast<int>(n3 > 0) + static_cast<int>(n4 > 0))
 {
   std::vector<const FiniteElement<dim, spacedim> *> fes;
   fes.push_back(&fe1);
@@ -269,7 +271,9 @@ FESystem<dim, spacedim>::FESystem(const FiniteElement<dim, spacedim> &fe1,
                                                        n4,
                                                        &fe5,
                                                        n5))
-  , base_elements((n1 > 0) + (n2 > 0) + (n3 > 0) + (n4 > 0) + (n5 > 0))
+  , base_elements(static_cast<int>(n1 > 0) + static_cast<int>(n2 > 0) +
+                  static_cast<int>(n3 > 0) + static_cast<int>(n4 > 0) +
+                  static_cast<int>(n5 > 0))
 {
   std::vector<const FiniteElement<dim, spacedim> *> fes;
   fes.push_back(&fe1);
